+++ cmd/agent_create.go
@@ -105,6 +105,7 @@ var agentCreateCmd = &cobra.Command{
 			),
 		)
 
+		basics.WithProgramOptions(ui.FormOptions()...)
 		if err := basics.Run(); err != nil {
 			return err
 		}